Allow custom install directory for Quickviewer
Quickviewer stores its configuration in a file located in the same directory as the executable. When installed to the default directory, configuration changes cannot be saved unless Quickviewer is run with elevated permissions.

This change allows for Quickviewer to be installed to a different directory (e.g. one that does not require elevated permissions). Backward compatibility with prior releases is maintained: if no directory is specified during installation, the default location will be used.